2010-04-14	Ralf Corsépius <ralf.corsepius@rtems.org>
	* libnetworking/sys/mbuf.h: Use uintptr_t instead of u_long
	for 16bit target compliance.

+++ cpukit/libnetworking/sys/mbuf.h
@@ -63,7 +63,7 @@
  */
 #define	mtod(m, t)	((t)((m)->m_data))
 #define	dtom(x)		((struct mbuf *)((intptr_t)(x) & ~(MSIZE-1)))
-#define	mtocl(x)	(((u_long)(x) - (u_long)mbutl) >> MCLSHIFT)
+#define	mtocl(x)	(((uintptr_t)(x) - (uintptr_t)mbutl) >> MCLSHIFT)
 #define	cltom(x)	((caddr_t)((u_long)mbutl + ((u_long)(x) << MCLSHIFT)))
 
 /*
